refactor to separate component, stop propagating radio button event + unit test
